+++ app/Http/Controllers/LivrosController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use App\Models\Livro;
 use Illuminate\Http\Request;
 use Session;
@@ -15,8 +16,12 @@ class LivrosController extends Controller
      */
     public function index()
     {
-        $livros = Livro::paginate(5);
-        return view('livro.index',array('livros' => $livros,'busca'=>null));
+        if (Auth::check()) {
+            $livros = Livro::simplepaginate(5);
+            return view('livro.index',array('livros' => $livros,'busca'=>null));
+        } else {
+            return redirect('login');
+        }
     }
 
     /**
@@ -25,9 +30,7 @@ class LivrosController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function buscar(Request $request) {
-        $livros = livro::where('titulo','LIKE','%'.$request->input('busca').'%')->orwhere
-        ('descricao','LIKE','%'.$request->input('busca').'%')->orwhere('autor','LIKE','%'.$request->input('busca').'%')
-        ->orwhere('editora','LIKE','%'.$request->input('busca').'%')->paginate(5);
+        $livros = livro::where('titulo','LIKE','%'.$request->input('busca').'%')->orwhere('descricao','LIKE','%'.$request->input('busca').'%')->orwhere('autor','LIKE','%'.$request->input('busca').'%')->orwhere('editora','LIKE','%'.$request->input('busca').'%')->simplepaginate(5);
         return view('livro.index',array('livros' => $livros,'busca'=>$request->input('busca')));
     }
 
@@ -39,7 +42,11 @@ class LivrosController extends Controller
      */
     public function create()
     {
-        return view('livro.create');
+        if (Auth::check()) {
+            return view('livro.create');
+        } else {
+            return redirect('login');
+        }
     }
 
     /**
@@ -50,20 +57,30 @@ class LivrosController extends Controller
      */
     public function store(Request $request)
     {
-    
-        $livro = new Livro();
-        $livro->titulo = $request->input('titulo');
-        $livro->descricao = $request->input('descricao');
-        $livro->autor = $request->input('autor');
-        $livro->editora = $request->input('editora');
-        $livro->ano = $request->input('ano');
-        if($livro->save()) {
-            if($request->hasFile('foto')){
-                $imagem = $request->file('foto');
-                $nomearquivo = md5($livro->id).".".$imagem->getClientOriginalExtension();
-                $request->file('foto')->move(public_path('.\img\livros'),$nomearquivo);
+        if (Auth::check()) {
+            $this->validate($request,[
+                'titulo' => 'required|min:3',
+                'descricao' => 'required',
+                'autor' => 'required',
+                'editora' => 'required',
+                'ano' => 'required',
+            ]);
+            $livro = new Livro();
+            $livro->titulo = $request->input('titulo');
+            $livro->descricao = $request->input('descricao');
+            $livro->autor = $request->input('autor');
+            $livro->editora = $request->input('editora');
+            $livro->ano = $request->input('ano');
+            if($livro->save()) {
+                if($request->hasFile('foto')){
+                    $imagem = $request->file('foto');
+                    $nomearquivo = md5($livro->id).".".$imagem->getClientOriginalExtension();
+                    $request->file('foto')->move(public_path('.\img\livros'),$nomearquivo);
+                }
+                return redirect('livros');
             }
-            return redirect('livros');
+        } else {
+            return redirect('login');
         }
     }
 
@@ -87,8 +104,12 @@ class LivrosController extends Controller
      */
     public function edit($id)
     {
-        $livro = Livro::find($id);
-        return view('livro.edit', array('livro' => $livro));
+        if (Auth::check()) {
+            $livro = Livro::find($id);
+            return view('livro.edit',array('livro' => $livro));
+        } else {
+            return redirect('login');
+        }
     }
 
     /**
@@ -100,34 +121,52 @@ class LivrosController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $livro = Livro::find($id);
-        if($request->hasFile('foto')){
-            $imagem = $request->file('foto');
-            $nomearquivo = md5($livro->id).".".$imagem->getClientOriginalExtension();
-            $request->file('foto')->move(public_path('.\img\livros'),$nomearquivo);
-        }
-        $livro->titulo = $request->input('titulo');
-        $livro->descricao = $request->input('descricao');
-        $livro->autor = $request->input('autor');
-        $livro->editora  = $request->input('editora');
-        $livro->ano = $request->input('ano');
-        if($livro->save()) {
-            Session::flash('mensagem','Livro alterado com sucesso');
-            return redirect('livros');
+        if (Auth::check()) {
+            $this->validate($request,[
+                'titulo' => 'required|min:3',
+                'descricao' => 'required',
+                'autor' => 'required',
+                'editora' => 'required',
+                'ano' => 'required',
+            ]);
+            $livro = Livro::find($id);
+            if($request->hasFile('foto')){
+                $imagem = $request->file('foto');
+                $nomearquivo = md5($livro->id).".".$imagem->getClientOriginalExtension();
+                $request->file('foto')->move(public_path('.\img\livros'),$nomearquivo);
+            }
+            $livro->titulo = $request->input('titulo');
+            $livro->descricao = $request->input('descricao');
+            $livro->autor = $request->input('autor');
+            $livro->editora  = $request->input('editora');
+            $livro->ano = $request->input('ano');
+            if($livro->save()) {
+                Session::flash('mensagem','Livro alterado com sucesso');
+                return redirect('livros');
+            }
+        } else {
+            return redirect('login');
         }
     }
 
     /**
      * Remove the specified resource from storage.
-     *
+     * @param \Illuminate\Http\Request $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $livro = Livro::find($id);
-        $livro->delete();
-        Session::flash('mensagem','Livro Excluído com Sucesso');
-        return redirect(url('livros/'));
+        if (Auth::check()) {
+            $livro = Livro::find($id);
+            if (isset($request->foto)) {
+            unlink($request->foto);
+            }
+            $livro->delete();
+            Session::flash('mensagem','Livro Excluído com Sucesso');
+            return redirect(url('livros/'));
+        } else {
+            return redirect('login');
+        }
     }
 }
